requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
arrow==1.3.0
binaryornot==0.4.4
cachetools==5.5.2
cashews==7.4.1
certifi==2025.7.14
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
distlib==0.4.0
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.18.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.1
multidict==6.6.3
nb-cli==1.4.2
nonebot-adapter-onebot==2.4.6
nonebot-plugin-apscheduler==0.5.0
nonebot-plugin-localstore==0.7.4
nonebot2==2.4.2
noneprompt==0.1.9
nonestorage==0.1.0
openai==1.98.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.3.4
pydantic_core==2.33.2
pyfiglet==1.0.3
Pygments==2.19.2
Pillow==10.4.0
pygtrie==2.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
rich==14.1.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
text-unidecode==1.3
tomlkit==0.13.3
tqdm==4.67.1
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.32.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==15.0.1
win32_setctime==1.2.0
yarl==1.20.1
redis==6.2.0
numpy
sentence-transformers
